Move changepoints_to_labels to utils
